+++ CHANGELOG.md
@@ -8,7 +8,13 @@
 * [antibodies] set missing mandatory fields in tissus edition
 * [mailer] remove default cc and mailerconfig cc option
 * [doc] fix install docs broken links
+* [core] in background process, reset core config parameters
+  with reset, core config params are loaded only once
 * [services] add close projects at invoice option
+
+### Other
+
+* [antibodies] add unit tests
 
 ## 2.9.1